test: fold the unconfigured-point tests into the transform census file
Both modules cover the same thing: what an adapter must do with a verdict it
cannot act on. Keeping them apart pushed this PR to 301 files, one past the
point where GitHub stops rendering the diff, and the split bought nothing.

No coverage is lost. The unconfigured-point classes move across whole and
reuse the census file's stubs rather than carrying their own near-duplicates.

Copilot-Session: 384bda36-161f-4689-9b2c-0a6d5c90776f
